Extended ref targets to include page numbers
Thanks to Melissa Jane Taylor for pointing out the issue in this
volume. Certainly applies elsewhere. I wrote a script to find and fix
these references (<ref> elements followed by the pattern “, p. #” or “,
pp. #”), and could apply this to other volumes. We could also add a
check for such references to our schema. We could also make it easier
to navigate from these page-based targets to their documents. Hopefully
this commit helps for now.