Add centralized frus-dates taxonomy for date analysis
- To make use of the date taxonomy, see the change to frus1977-80v28.
We just need to reference the taxonomy in the TEI Header using XInclude.
- Also, I’ve added the taxonomy elements to our ODD, enumerated the
current values (making entry easier since oXygen will auto-suggest
these values), and I added a 2nd check in the schematron to ensure that
the ana values are defined in the TEI header (either explicitly there
or via xinclude); we should use the XInclude method
- As we add more values to the taxonomy, we will need to add register
them both in the frus-dates.xml file and in the ODD and regenerate the
frus.rnc file.
- For more on XInclude, see “Including Document Parts with XInclude”
from
https://www.oxygenxml.com/doc/versions/18.1/ug-editor/topics/including-d
ocument-parts-with-XInclude.html
- Found techniques in conversations on TEI-L. Special thanks to the
Shelley-Godwin archive’s sample at
https://github.com/umd-mith/sga/blob/master/data/tei/ox/ox-ms_abinger_c5
6.xml#L202.