legacy review volume delivery from vendor
- Before committing, merged @WaxCylinderRevival’s dateline fixes. Also
fixed an issue with a couple of these datelines in which the month was
tagged as placeName.
- Ran schematron checks. Expected errors not fixed in this commit
include: ambiguous cross-references and list element within a table
- Unexpected errors fixed in this commit: an “ibid” volume cross
reference pointed to the wrong volume; traced the ibid reference back
and corrected the various references. In one case this led me to supply
a new errata statement fixing an error in the original volume (see
frus1964-68v07/errata).