corrected footnote refs
@awmarrs noticed that Johnson admin vols occasionally
include multiple references to the same footnote in the document text,
but only print the footnote once. This practice threw off our footnote
numbering and I corrected it to include two references to footnote 2 in
d34. The footnote displays twice, so @joewiz please let me know if this
is okay or if there is a better way to solve this problem?