q4 2017 ongoing
empty head elements fixed in 240 docs

put <head>[Untitled]</head> in docs with no head content, also used
instead of source note or seg tag in cases of [Extract], [Inclosure],
[Translation], etc.

Inserted dateline and closers in 94+ docs with placeName and date tags.

Question: 22 instances of “Done at”—should I treat this as part of a
closer though it’s usually a full sentence or paragraph, as it does
contain the placeName and date information? Example: "Done at
Washington the 21st day of February, in the nineteen hundred and
eleventh year of the Christian era, corresponding to the 21st day of
the 2d month of the 44th year of Meiji.” (p. 320)

checklist complete. format and indent applied. close #147