Add query to renumber doc & note xml:ids
Renumbers all @xml:id and @n values of document divs sequentially from 1 and updates note @xml:id values accordingly.

This is designed to be run from an oXygen Transformation Scenario (see the forthcoming commit to hsg-project), with saxon-configuration.xml set as the Saxon configuration file.

oXygen produces a copy of the file with the .bak file extension - despite my attempts to prevent this. So when using this be sure to delete the .bak file.