Q3 volumes final checks
@waxcylinderrevival turning these over for dates project checks and inclusion! Remaining six volumes completed by tomorrow at latest. Transforms and format and ident applied in all.

Notes: doc numbering off in 1872p2v5 at d63 and renumbering script didn't work for me--not sure if that's a problem on my end or because one div is a document and the other a toc?

Note within note error still present in parallel translations in 1872p2v4.

Let me know if I missed (or accidentally introduced) any errors!