Adapt source note schematron checks for pre-1955
If the source note is the 1st element in the document div, it's a pre-1955 document, before the "Source:" convention was present/consistent, so we won’t raise an error. Also common in our FRUS TEI encoding is to encode bracketed notes like [Extract] as <note type=“source”>, so until we further distinguish between Source Notes and other inline editorial interventions, this check is less useful in the pre-1955 era.