Fix div nesting (move docs into correct comp)
As discovered by XPath via oXygen’s XPath toolbar, with the frus volumes directory selected:

//div[@type eq "compilation"]/following-sibling::div[@type="document"][1]

And an expanded review of all docs that are siblings with compilations, chapters, or subchapters:

//div[@type = ("compilation", "chapter", "subchapter")]/(preceding-sibling::div[@type="document"][1] | following-sibling::div[@type="document"][1])

Note: This XPath still returns 252 hits, indicating chapters and subchapters interleaved with documents, but in these remaining cases, we have correctly reproduced the structure of the original volume. For example, notice the Great Britain chapter in https://history.state.gov/historicaldocuments/frus1864p1 contains subchapters interleaved with documents. A much more common case is an editorial note at the start of a compilation, followed by chapters.

The purpose of this commit was just to correct cases where our encoding diverged from the volume’s structure.

The commit also contains a new schema check to warn when such cases are found.