A partial pass, fixing en dashes
This would be nice to catch via Schematron, but I’m not sure it’s possible. One indicator is the pattern (case-sensitive regex):
[a-z]–[a-z]
This catches many cases of en dashes that should be hyphens or em dashes, including Chou En–lai.
This regex guided most of the changes in this commit. It would be nice to find a fuller means to test this & related phenomena.