Add subtypes from XSL enrichment scripts
... and separate div type & subtype checks into their own patterns so that both rules run. This separation is needed, because according to the timeless Schematron tutorial from Mulberry's Piez and Lapere (https://www.mulberrytech.com/papers/schematron-Philly.pdf p. 13):

"In a pattern, (a) the first rule with a @context matching the element is applied, and (b) subsequent rules in the same pattern that match the element are ignored. In the rule, all assertions are tested with the matched element as context."

I also checked all of our rules' @context values to be sure the others didn't inadvertently cause missed schema checks as this one had managed to do.