Fix whitespace around <hi> element with leading...
or trailing punctuation, focusing on front matter lists but including many instances in the body. Such whitespace is significant in mixed content, and can produce an extra space between, say a person's name and a colon (in a transcript) or comma or period (in a sentence). Neglecting whitespace can also cause a space between words to be omitted.